add canvas reaction, rework server response

+++ src/main/java/com/vuhtang/lab2/utils/Shot.java
@@ -2,5 +2,5 @@ package com.vuhtang.lab2.utils;
 
 import java.util.Date;
 
-public record Shot(Integer x, Double y, Double r, Date currTime, Long execTime, String result) {
+public record Shot(Double x, Double y, Double r, Date currTime, Long execTime, String result) {
 }
