+++ src/main/java/com/vuhtang/lab2/utils/ShotChecker.java
@@ -1,5 +1,5 @@
 package com.vuhtang.lab2.utils;
 
 public interface ShotChecker {
-    Shot takeShot(Integer x, Double y, Double r);
+    Shot takeShot(Double x, Double y, Double r);
 }
